refactoring and expanding About page + route
